flatpak: Make home folder access read-only
The goal is for asma to eventually not have any access to the filesystem
when in a flatpak, and to make use of portals to find the Arma 3 folder.
This will not be a problem if/when we add a config file, as flatpak apps
get their own folders for this (as long as we stick to the xdg dir
standard).