settings: Switch file chooser to a native dialog
This allows gtk to pick the best file chooser for the situation. In our
case, it lets the flatpak FileChooser portal be used. Unfortunately, the
flatpak FileChooser portal does not allow opening folders, which we
need, so this doesn't really change anything. If there is ever a way to
allow folders through, then we should have everything in place already.

Unfortunately, the lack of folder support also means that we will need
to have read only access to the entire host filesystem, as we have no
way of allowing the user to specify through a portal where the Arma 3
folder is.