Delete the flatpak manifest
While I have learned a lot from this experiment, I do not have
intentions to take it further right now. There is still quite a bit of
work to be done before asma works properly in a sandbox. If anyone else
wants to improve upon this, it will be available in a separate branch
or at least the history of this repo.