First rewrite commit
This is nonfunctional, as I am replacing almost everything in a very
destructive manner. Hence, my commits will not be as atomic as usual.
Sorry!

Right now, we have a couple xml files which define the layout. These are
packed into the binary with gresource, and allow the C code to mostly
consist of logic instead.

What little I have rewritten is close to the Linux kernel C coding
style. It uses Gnome libraries wherever possible. The window(s) will
conform more closely to the Gnome Human Interface Guidelines.